Rename 'parsing-messages' file
