+++ src/component/parser/BpmnParser.ts
@@ -17,7 +17,7 @@
 import BpmnModel from '../../model/bpmn/internal/BpmnModel';
 import BpmnXmlParser from './xml/BpmnXmlParser';
 import BpmnJsonParser, { newBpmnJsonParser } from './json/BpmnJsonParser';
-import { ParsingMessageCollector } from './parsing-messages-management';
+import { ParsingMessageCollector } from './parsing-messages';
 
 /**
  * @internal
